requirements.txt
# AI量化交易系统 - Python依赖

# 核心数据处理
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1

# 数据获取
akshare==1.11.60
tushare==1.3.4
baostock==0.8.9

# 机器学习/深度学习
torch==2.0.1
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==4.0.0
catboost==1.2

# 量化回测
backtrader==1.9.78.123
empyrical==0.5.5
pyfolio==0.9.2

# 技术指标
ta-lib==0.4.28
pandas-ta==0.3.14b

# 数据可视化
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# 数据库
pymysql==1.1.0
sqlalchemy==2.0.19
redis==4.6.0

# 配置管理
pyyaml==6.0.1
python-dotenv==1.0.0

# 日志
loguru==0.7.0

# API和网络请求
requests==2.31.0
aiohttp==3.8.5
websockets==11.0.3

# 网页自动化
selenium==4.15.2
webdriver-manager==4.0.1

# 桌面自动化
pyautogui==0.9.54
psutil==5.9.5
pillow==10.0.0  # pyautogui依赖

# 工具库
python-dateutil==2.8.2
pytz==2023.3
schedule==1.2.0

# 消息通知
wxpy==0.3.9.8  # 微信通知（可选）

# 开发工具
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0
mypy==1.4.1

# 文档生成
sphinx==7.1.0
sphinx-rtd-theme==1.2.2
